IntelGPU backend: Separate backend and executable classes
