+++ src/ngraph/runtime/intelgpu/intelgpu_backend.hpp
@@ -20,7 +20,6 @@
 #include <memory>
 
 #include <CPP/engine.hpp>
-#include <CPP/network.hpp>
 
 #include "ngraph/runtime/backend.hpp"
 
@@ -31,7 +30,6 @@ namespace ngraph
         namespace intelgpu
         {
             class IntelGPUBackend;
-            class IntelGPUExecutable;
         }
     }
 }
@@ -66,40 +64,4 @@ private:
     bool m_function_cache_disabled = false;
     bool m_disable_backend_optimizations = false;
     std::string m_cldnn_dump_dir = std::string("intelgpu_codegen");
-};
-
-class ngraph::runtime::intelgpu::IntelGPUExecutable : public runtime::Executable
-{
-public:
-    IntelGPUExecutable(std::shared_ptr<Function> func,
-                       std::shared_ptr<cldnn::network> network,
-                       bool enable_timing,
-                       bool enable_profile,
-                       double compilation_time,
-                       double consumed_memory,
-                       size_t profile_lines_limit_count);
-
-    bool call(const std::vector<std::shared_ptr<runtime::Tensor>>& outputs,
-              const std::vector<std::shared_ptr<runtime::Tensor>>& inputs) override;
-
-    std::vector<PerformanceCounter> get_performance_data() const override;
-
-private:
-    std::shared_ptr<Function> m_function;
-    std::shared_ptr<cldnn::network> m_cldnn_network = nullptr;
-    bool m_performance_counters_enabled = false;
-    bool m_profile_enable = false;
-    double m_compilation_time = 0.0;
-    double m_consumed_memory = 0.0;
-    long m_profile_lines_limit_count = 10;
-    std::string delim = std::string(":");
-
-    // Statistic related things
-    void print_call_performance(const std::shared_ptr<cldnn::network> network,
-                                const std::shared_ptr<Function> func,
-                                double time_compile,
-                                double time_call,
-                                double mem_compilation_consumed,
-                                double mem_call_consumed,
-                                double mem_current) const;
 };
